Update the project. The connexion process is updated
Added features:
If the user is logged, the app will try to connect it.
On the ProjectActivity. If swip down it reload the ProjectList
You no longer can't click on an "Page" it this page is already display.
The app now have internal storage, to display the login, store the token given by the server.
...